+++ components/Checkbox.js
@@ -6,7 +6,7 @@ export default function Checkbox(props) {
     const { toggleAgree, label } = props
   return (
     <label class="checkbox">
-      <p class="has-text-justified py-5">
+      <p class="has-text-justified has-text-centered py-5">
         <input type="checkbox" onClick={toggleAgree} />
         {label}
       </p>