+++ components/SectionTitle.js
@@ -4,7 +4,7 @@ export default function SectionTitle(props) {
     const {name} = props
   return (
     <div class='container my-6'>
-      <h1 className="title is-size-1 my-6">{name}</h1>
+      <h3 className="title is-size-3 my-6">{name}</h3>
     </div>
   )
 }