+++ pages/info/service/index.js
@@ -1,0 +1,37 @@
+export const faqs = [{
+    question: "What is the Lorem Health Solutions?",
+    answer: "Dolor Sit Amet."
+  },
+  {
+    question: "Is there a membership fee?",
+    answer: "Sed ut perspiciatis unde omnis"
+  },
+  {
+    question: "What is processing of personal information?",
+    answer: "iste natus error sit voluptatem accusantium doloremque laudantium."
+  },
+  {
+    question: "How long is the membership lasts?",
+    answer: "totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et."
+  },
+  {
+    question: "Who can see personal details?",
+    answer: "quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit."
+  },
+  {
+    question: "How can I contact lorem helth solutions?",
+    answer: "sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam"
+  },
+  {
+    question: "How does my data will be used?",
+    answer: "qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem."
+  },
+  {
+    question: "Is there an ID needed to register?",
+    answer: "Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur."
+  },
+  {
+    question: "How long should I have to wait to know if I'm approved or not?",
+    answer: "Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla."
+  },
+]