+++ pages/info/services.js
@@ -11,13 +11,6 @@ export default function services() {
         <div className={styles.services_hero}>
           <h1 class="is-size-1 has-text-centered has-text-weight-bold">Services</h1>
         </div>
-        <div class="container is-fluid my-5">
-          <p class="has-text-centered my-5">
-            Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet,
-            consectetur, adipisci velit, sed quia non numquam eius modi tempora
-            incidunt ut labore et dolore magnam aliquam quaerat voluptatem
-          </p>
-        </div>
         <div class="container is-fluid mb-5">
           <div class="columns is-desktop is-danger">
             <div class="column">